pyproject.toml
[project]
name = "iot-rag"
version = "0.1.0"
description = "A Retrieval-Augmented Generation (RAG) system for Internet of Things (IoT) applications."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "arize-phoenix-otel>=0.13.1",
  "chromadb>=1.2.1",
  "langchain>=1.0.2",
  "langchain-core>=1.0.1",
  "langchain-google-genai>=3.0.0",
  "langchain-text-splitters>=1.0.0",
  "openinference-instrumentation-langchain>=0.1.54",
  "pymupdf>=1.26.5",
  "python-dotenv>=1.1.1",
]

[project.scripts]
agent = "agent.cli:main"
rag = "rag.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag", "src/agent", "src/evaluation"]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.14.2",
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "N", "W"]
